contract rename
