+++ contracts/concentrator/strategies/AutoCompoundingConvexCurveStrategy.sol
@@ -13,12 +13,12 @@ import "./AutoCompoundingStrategyBase.sol";
 
 // solhint-disable no-empty-blocks
 
-contract ConvexCurveAutoCompoundingStrategy is AutoCompoundingStrategyBase {
+contract AutoCompoundingConvexCurveStrategy is AutoCompoundingStrategyBase {
   using SafeERC20 for IERC20;
 
   /// @inheritdoc IConcentratorStrategy
   // solhint-disable const-name-snakecase
-  string public constant override name = "ConvexCurveAutoCompounding";
+  string public constant override name = "AutoCompoundingConvexCurve";
 
   /// @dev The address of Convex Booster.
   address private constant BOOSTER = 0xF403C135812408BFbE8713b5A23a04b3D48AAE31;
