+++ contracts/concentrator/strategies/ManualCompoundingCurveGaugeStrategy.sol
@@ -5,8 +5,8 @@ pragma solidity ^0.7.6;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
 
-import "../../interfaces/IConvexBasicRewards.sol";
-import "../../interfaces/IConvexBooster.sol";
+import "../../interfaces/ICurveGauge.sol";
+import "../../interfaces/ICurveTokenMinter.sol";
 import "../../interfaces/IZap.sol";
 
 import "./ManualCompoundingStrategyBase.sol";
@@ -14,51 +14,50 @@ import "./ManualCompoundingStrategyBase.sol";
 // solhint-disable no-empty-blocks
 // solhint-disable reason-string
 
-contract ConvexCurveStrategy is ManualCompoundingStrategyBase {
+contract ManualCompoundingCurveGaugeStrategy is ManualCompoundingStrategyBase {
   using SafeERC20 for IERC20;
 
   /// @inheritdoc IConcentratorStrategy
   // solhint-disable const-name-snakecase
-  string public constant override name = "ConvexCurve";
+  string public constant override name = "ManualCompoundingCurveGauge";
 
-  /// @dev The address of Convex Booster.
-  address private constant BOOSTER = 0xF403C135812408BFbE8713b5A23a04b3D48AAE31;
+  /// @dev The address of CRV token.
+  address private constant CRV = 0xD533a949740bb3306d119CC777fa900bA034cd52;
 
-  /// @notice The pid of Convex reward pool.
-  uint256 public pid;
+  /// @dev The address of CRV minter.
+  address private constant MINTER = 0xd061D61a4d941c39E5453435B6345Dc261C2fcE0;
 
   /// @notice The address of staking token.
   address public token;
 
-  /// @notice The address of Convex rewards contract.
-  address public rewarder;
+  /// @notice The address of Curve gauge contract.
+  address public gauge;
 
   function initialize(
     address _operator,
     address _token,
-    address _rewarder,
+    address _gauge,
     address[] memory _rewards
   ) external initializer {
     ConcentratorStrategyBase._initialize(_operator, _rewards);
 
-    IERC20(_token).safeApprove(BOOSTER, uint256(-1));
+    IERC20(_token).safeApprove(_gauge, uint256(-1));
 
-    pid = IConvexBasicRewards(_rewarder).pid();
     token = _token;
-    rewarder = _rewarder;
+    gauge = _gauge;
   }
 
   /// @inheritdoc IConcentratorStrategy
   function deposit(address, uint256 _amount) external override onlyOperator {
     if (_amount > 0) {
-      IConvexBooster(BOOSTER).deposit(pid, _amount, true);
+      ICurveGauge(gauge).deposit(_amount);
     }
   }
 
   /// @inheritdoc IConcentratorStrategy
   function withdraw(address _recipient, uint256 _amount) external override onlyOperator {
     if (_amount > 0) {
-      IConvexBasicRewards(rewarder).withdrawAndUnwrap(_amount, false);
+      ICurveGauge(gauge).withdraw(_amount);
       IERC20(token).safeTransfer(_recipient, _amount);
     }
   }
@@ -71,7 +70,10 @@ contract ConvexCurveStrategy is ManualCompoundingStrategyBase {
     for (uint256 i = 0; i < rewards.length; i++) {
       _amounts[i] = IERC20(_rewards[i]).balanceOf(address(this));
     }
-    IConvexBasicRewards(rewarder).getReward();
+    address _gauge = gauge;
+    ICurveTokenMinter(MINTER).mint(_gauge);
+    // some gauge has no extra rewards
+    try ICurveGauge(_gauge).claim_rewards() {} catch {}
     for (uint256 i = 0; i < rewards.length; i++) {
       _amounts[i] = IERC20(_rewards[i]).balanceOf(address(this)) - _amounts[i];
     }
