+++ test/concentrator/strategies/ManualCompoundingCurveGaugeStrategy.spec.ts
@@ -5,7 +5,7 @@ import { expect } from "chai";
 import { BigNumber, constants } from "ethers";
 import { ethers, network } from "hardhat";
 import { TOKENS, VAULT_CONFIG, ZAP_ROUTES } from "../../../scripts/utils";
-import { AladdinZap, ConvexCurveAutoCompoundingStrategy, IConvexBasicRewards, MockERC20 } from "../../../typechain";
+import { AladdinZap, ManualCompoundingCurveGaugeStrategy, ICurveGauge, MockERC20 } from "../../../typechain";
 import { request_fork } from "../../utils";
 
 const UNDERLYING: {
@@ -14,11 +14,11 @@ const UNDERLYING: {
     deployer: string;
     token: string;
     pool: string;
-    rewarder: string;
+    gauge: string;
     holder: string;
     amount: string;
     rewards: string[];
-    intermediate: string;
+    intermediates: string[];
   };
 } = {
   frxeth: {
@@ -26,64 +26,75 @@ const UNDERLYING: {
     deployer: "0xDA9dfA130Df4dE4673b89022EE50ff26f6EA73Cf",
     token: "0xf43211935C781D5ca1a41d2041F397B8A7366C7A",
     pool: "0xa1F8A6807c402E4A15ef4EBa36528A3FED24E577",
-    rewarder: "0xbD5445402B0a287cbC77cb67B2a52e2FC635dce4",
+    gauge: "0x2932a86df44Fe8D2A706d8e9c5d51c24883423F5",
     holder: "0xadd85e4abbb426e895f35e0a2576e22a9bbb7a57",
     amount: "10",
     rewards: ["CVX", "CRV"],
-    intermediate: "WETH",
+    intermediates: ["CRV", "FXS", "WETH"],
   },
   steth: {
     fork: 16124420,
     deployer: "0xDA9dfA130Df4dE4673b89022EE50ff26f6EA73Cf",
     token: "0x06325440D014e39736583c165C2963BA99fAf14E",
     pool: "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022",
-    rewarder: "0x0A760466E1B4621579a82a39CB56Dda2F4E70f03",
+    gauge: "0x182B723a58739a9c974cFDB385ceaDb237453c28",
     holder: "0x13e382dfe53207E9ce2eeEab330F69da2794179E",
     amount: "100",
     rewards: ["CVX", "CRV", "LDO"],
-    intermediate: "WETH",
+    intermediates: ["CRV", "WETH"],
   },
   cbeth: {
     fork: 16124420,
     deployer: "0xDA9dfA130Df4dE4673b89022EE50ff26f6EA73Cf",
     token: "0x5b6C539b224014A09B3388e51CaAA8e354c959C8",
     pool: "0x5FAE7E604FC3e24fd43A72867ceBaC94c65b404A",
-    rewarder: "0x5d02EcD9B83f1187e92aD5be3d1bd2915CA03699",
+    gauge: "0xAd96E10123Fa34a01cf2314C42D75150849C9295",
     holder: "0xe4d7e7b90519445585635c3b383c9d86c0596e57",
     amount: "10",
     rewards: ["CVX", "CRV"],
-    intermediate: "WETH",
+    intermediates: ["CRV", "WETH"],
+  },
+  cvxfxs: {
+    fork: 16124420,
+    deployer: "0xDA9dfA130Df4dE4673b89022EE50ff26f6EA73Cf",
+    token: "0xF3A43307DcAFa93275993862Aae628fCB50dC768",
+    pool: "0xd658A338613198204DCa1143Ac3F01A722b5d94A",
+    gauge: "0xab1927160ec7414c6fa71763e2a9f3d107c126dd",
+    holder: "0xdc88d12721f9ca1404e9e6e6389ae0abdd54fc6c",
+    amount: "1000",
+    rewards: ["CVX", "CRV", "FXS"],
+    intermediates: ["CRV", "WETH", "FXS"],
   },
 };
 
-describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
+describe("ManualCompoundingCurveGaugeStrategy.spec", async () => {
   context("auth", async () => {
     let deployer: SignerWithAddress;
     let operator: SignerWithAddress;
-    let strategy: ConvexCurveAutoCompoundingStrategy;
+    let strategy: ManualCompoundingCurveGaugeStrategy;
     let token: MockERC20;
 
     beforeEach(async () => {
       [deployer, operator] = await ethers.getSigners();
 
-      const ConvexCurveAutoCompoundingStrategy = await ethers.getContractFactory(
-        "ConvexCurveAutoCompoundingStrategy",
+      const ManualCompoundingCurveGaugeStrategy = await ethers.getContractFactory(
+        "ManualCompoundingCurveGaugeStrategy",
         deployer
       );
-      strategy = await ConvexCurveAutoCompoundingStrategy.deploy();
+      strategy = await ManualCompoundingCurveGaugeStrategy.deploy();
       await strategy.deployed();
 
       const MockERC20 = await ethers.getContractFactory("MockERC20", deployer);
       token = await MockERC20.deploy("x", "y", 18);
       await token.deployed();
 
-      const MockConvexBasicRewards = await ethers.getContractFactory("MockConvexBasicRewards", deployer);
-      const rewarder = await MockConvexBasicRewards.deploy(1, token.address);
-      await rewarder.deployed();
+      const MockCurveGaugeV4V5 = await ethers.getContractFactory("MockCurveGaugeV4V5", deployer);
+      const gauge = await MockCurveGaugeV4V5.deploy();
+      await gauge.deployed();
 
-      await strategy.initialize(operator.address, token.address, rewarder.address, []);
+      await strategy.initialize(operator.address, token.address, gauge.address, []);
 
-      expect(await strategy.name()).to.eq("ConvexCurveAutoCompounding");
+      expect(await strategy.name()).to.eq("ManualCompoundingCurveGauge");
     });
 
     it("should revert, when initialize again", async () => {
@@ -161,10 +172,6 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
           "ConcentratorStrategy: only operator"
         );
       });
-
-      it("should succeed, when operator call harvest", async () => {
-        await strategy.connect(operator).harvest(deployer.address, token.address);
-      });
     });
 
     context("#prepareMigrate", async () => {
@@ -195,19 +202,19 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
       deployer: string;
       token: string;
       pool: string;
-      rewarder: string;
+      gauge: string;
       holder: string;
       amount: string;
       rewards: string[];
-      intermediate: string;
+      intermediates: string[];
     }
   ) => {
     context(`${name}`, async () => {
       let deployer: SignerWithAddress;
       let holder: SignerWithAddress;
       let zap: AladdinZap;
-      let strategy: ConvexCurveAutoCompoundingStrategy;
-      let rewarder: IConvexBasicRewards;
+      let strategy: ManualCompoundingCurveGaugeStrategy;
+      let gauge: ICurveGauge;
 
       beforeEach(async () => {
         request_fork(config.fork, [config.deployer, config.holder]);
@@ -215,7 +222,7 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
         holder = await ethers.getSigner(config.holder);
         await deployer.sendTransaction({ to: holder.address, value: ethers.utils.parseEther("10") });
 
-        rewarder = await ethers.getContractAt("IConvexBasicRewards", config.rewarder, deployer);
+        gauge = await ethers.getContractAt("ICurveGauge", config.gauge, deployer);
 
         const AladdinZap = await ethers.getContractFactory("AladdinZap", deployer);
         zap = await AladdinZap.deploy();
@@ -227,25 +234,18 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
         for (const [symbol, routes] of Object.entries(VAULT_CONFIG[name].deposit)) {
           await zap.updateRoute(TOKENS[symbol].address, config.token, routes);
         }
-        for (const reward of config.rewards) {
-          await zap.updateRoute(
-            TOKENS[reward].address,
-            TOKENS[config.intermediate].address,
-            ZAP_ROUTES[reward][config.intermediate]
-          );
-        }
 
-        const ConvexCurveAutoCompoundingStrategy = await ethers.getContractFactory(
-          "ConvexCurveAutoCompoundingStrategy",
+        const ManualCompoundingCurveGaugeStrategy = await ethers.getContractFactory(
+          "ManualCompoundingCurveGaugeStrategy",
           deployer
         );
-        strategy = await ConvexCurveAutoCompoundingStrategy.deploy();
+        strategy = await ManualCompoundingCurveGaugeStrategy.deploy();
         await strategy.deployed();
 
         await strategy.initialize(
           deployer.address,
           config.token,
-          config.rewarder,
+          config.gauge,
           config.rewards.map((symbol) => TOKENS[symbol].address)
         );
       });
@@ -255,7 +255,7 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
         const amount = ethers.utils.parseEther(config.amount);
         await token.transfer(strategy.address, amount);
         await strategy.deposit(deployer.address, amount);
-        expect(await rewarder.balanceOf(strategy.address)).to.eq(amount);
+        expect(await gauge.balanceOf(strategy.address)).to.eq(amount);
       });
 
       it("should succeed when withdraw", async () => {
@@ -263,32 +263,83 @@ describe("ConvexCurveAutoCompoundingStrategy.spec", async () => {
         const amount = ethers.utils.parseEther(config.amount);
         await token.transfer(strategy.address, amount);
         await strategy.deposit(deployer.address, amount);
-        expect(await rewarder.balanceOf(strategy.address)).to.eq(amount);
+        expect(await gauge.balanceOf(strategy.address)).to.eq(amount);
 
         const before = await token.balanceOf(deployer.address);
         await strategy.withdraw(deployer.address, amount);
         const after = await token.balanceOf(deployer.address);
         expect(after.sub(before)).to.eq(amount);
-        expect(await rewarder.balanceOf(strategy.address)).to.eq(constants.Zero);
+        expect(await gauge.balanceOf(strategy.address)).to.eq(constants.Zero);
       });
 
-      it("should succeed when harvest", async () => {
-        const token = await ethers.getContractAt("MockERC20", config.token, holder);
-        const amount = ethers.utils.parseEther(config.amount);
-        await token.transfer(strategy.address, amount);
-        await strategy.deposit(deployer.address, amount);
-        expect(await rewarder.balanceOf(strategy.address)).to.eq(amount);
+      for (const intermediate of config.intermediates) {
+        it(`should succeed when harvest to ${intermediate}`, async () => {
+          for (const reward of config.rewards) {
+            if (intermediate === reward) continue;
+            await zap.updateRoute(
+              TOKENS[reward].address,
+              TOKENS[intermediate].address,
+              ZAP_ROUTES[reward][intermediate]
+            );
+          }
 
-        const timestamp = (await ethers.provider.getBlock("latest")).timestamp;
-        // make sure 7 days passed, then the rewards will not increase anymore.
-        await network.provider.send("evm_setNextBlockTimestamp", [timestamp + 86400 * 7]);
-        await network.provider.send("evm_mine");
+          const token = await ethers.getContractAt("MockERC20", config.token, holder);
+          const amount = ethers.utils.parseEther(config.amount);
+          await token.transfer(strategy.address, amount);
+          await strategy.deposit(deployer.address, amount);
+          expect(await gauge.balanceOf(strategy.address)).to.eq(amount);
 
-        const harvested = await strategy.callStatic.harvest(zap.address, TOKENS[config.intermediate].address);
-        expect(harvested).to.gt(constants.Zero);
-        await strategy.harvest(zap.address, TOKENS[config.intermediate].address);
-        expect(await rewarder.balanceOf(strategy.address)).to.eq(amount.add(harvested));
-      });
+          const timestamp = (await ethers.provider.getBlock("latest")).timestamp;
+          // make sure 7 days passed, then the rewards will not increase anymore.
+          await network.provider.send("evm_setNextBlockTimestamp", [timestamp + 86400 * 7]);
+          await network.provider.send("evm_mine");
+
+          const intermediateToken = await ethers.getContractAt("MockERC20", TOKENS[intermediate].address, holder);
+          const harvested = await strategy.callStatic.harvest(zap.address, TOKENS[intermediate].address);
+          expect(harvested).to.gt(constants.Zero);
+
+          const before = await intermediateToken.balanceOf(deployer.address);
+          await strategy.harvest(zap.address, TOKENS[intermediate].address);
+          const after = await intermediateToken.balanceOf(deployer.address);
+          expect(after.sub(before)).to.closeToBn(harvested, harvested.div(1e4));
+        });
+
+        if (intermediate === "WETH") {
+          it(`should succeed when harvest to ETH`, async () => {
+            for (const reward of config.rewards) {
+              if (intermediate === reward) continue;
+              await zap.updateRoute(
+                TOKENS[reward].address,
+                TOKENS[intermediate].address,
+                ZAP_ROUTES[reward][intermediate]
+              );
+            }
+
+            const token = await ethers.getContractAt("MockERC20", config.token, holder);
+            const amount = ethers.utils.parseEther(config.amount);
+            await token.transfer(strategy.address, amount);
+            await strategy.deposit(deployer.address, amount);
+            expect(await gauge.balanceOf(strategy.address)).to.eq(amount);
+
+            const timestamp = (await ethers.provider.getBlock("latest")).timestamp;
+            // make sure 7 days passed, then the rewards will not increase anymore.
+            await network.provider.send("evm_setNextBlockTimestamp", [timestamp + 86400 * 7]);
+            await network.provider.send("evm_mine");
+
+            const harvested = await strategy.callStatic.harvest(zap.address, constants.AddressZero);
+            expect(harvested).to.gt(constants.Zero);
+
+            const before = await deployer.getBalance();
+            const tx = await strategy.harvest(zap.address, constants.AddressZero);
+            const receipt = await tx.wait();
+            const after = await deployer.getBalance();
+            expect(after.sub(before).add(receipt.gasUsed.mul(receipt.effectiveGasPrice))).to.closeToBn(
+              harvested,
+              harvested.div(1e4)
+            );
+          });
+        }
+      }
     });
   };
